use `.combineLatest()` instead of `.merge()` in model.js
It seemed that the issue with merge() was that there was a
race condition after the latest changes to @cycle/storage
and it was not guaranteed that the two streams were actually
properly combinable with scan. I am now using combineLatest which
makes sure that you always have two items of the two streams to combine.